Move inMemory iterators back to utils as in ASF
