+++ src/java/org/apache/cassandra/index/sai/utils/InMemoryPartitionIterator.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.cassandra.index.sai.iterators;
+package org.apache.cassandra.index.sai.utils;
 
 import java.util.Iterator;
 import java.util.Map;
@@ -24,23 +24,20 @@ import java.util.SortedMap;
 import java.util.TreeSet;
 
 import org.apache.cassandra.db.DecoratedKey;
-import org.apache.cassandra.db.DeletionTime;
 import org.apache.cassandra.db.ReadCommand;
 import org.apache.cassandra.db.RegularAndStaticColumns;
-import org.apache.cassandra.db.partitions.UnfilteredPartitionIterator;
-import org.apache.cassandra.db.rows.EncodingStats;
+import org.apache.cassandra.db.partitions.PartitionIterator;
 import org.apache.cassandra.db.rows.Row;
+import org.apache.cassandra.db.rows.RowIterator;
 import org.apache.cassandra.db.rows.Unfiltered;
-import org.apache.cassandra.db.rows.UnfilteredRowIterator;
-import org.apache.cassandra.index.sai.utils.PartitionInfo;
 import org.apache.cassandra.schema.TableMetadata;
 
-public class InMemoryUnfilteredPartitionIterator implements UnfilteredPartitionIterator
+public class InMemoryPartitionIterator implements PartitionIterator
 {
     private final ReadCommand command;
     private final Iterator<Map.Entry<PartitionInfo, TreeSet<Unfiltered>>> partitions;
 
-    public InMemoryUnfilteredPartitionIterator(ReadCommand command, SortedMap<PartitionInfo, TreeSet<Unfiltered>> rowsByPartitions)
+    public InMemoryPartitionIterator(ReadCommand command, SortedMap<PartitionInfo, TreeSet<Unfiltered>> rowsByPartitions)
     {
         this.command = command;
         this.partitions = rowsByPartitions.entrySet().iterator();
@@ -58,27 +55,21 @@ public class InMemoryUnfilteredPartitionIterator implements UnfilteredPartitionI
     }
 
     @Override
-    public UnfilteredRowIterator next()
+    public RowIterator next()
     {
-        return new InMemoryUnfilteredRowIterator(partitions.next());
-    }
-
-    @Override
-    public TableMetadata metadata()
-    {
-        return command.metadata();
+        return new InMemoryRowIterator(partitions.next());
     }
 
 
-    private class InMemoryUnfilteredRowIterator implements UnfilteredRowIterator
+    private class InMemoryRowIterator implements RowIterator
     {
         private final PartitionInfo partitionInfo;
-        private final Iterator<Unfiltered> unfiltereds;
+        private final Iterator<Unfiltered> rows;
 
-        public InMemoryUnfilteredRowIterator(Map.Entry<PartitionInfo, TreeSet<Unfiltered>> partition)
+        public InMemoryRowIterator(Map.Entry<PartitionInfo, TreeSet<Unfiltered>> rows)
         {
-            this.partitionInfo = partition.getKey();
-            this.unfiltereds = partition.getValue().iterator();
+            this.partitionInfo = rows.getKey();
+            this.rows = rows.getValue().iterator();
         }
 
         @Override
@@ -89,13 +80,13 @@ public class InMemoryUnfilteredPartitionIterator implements UnfilteredPartitionI
         @Override
         public boolean hasNext()
         {
-            return unfiltereds.hasNext();
+            return rows.hasNext();
         }
 
         @Override
-        public Unfiltered next()
+        public Row next()
         {
-            return unfiltereds.next();
+            return (Row) rows.next();
         }
 
         @Override
@@ -126,18 +117,6 @@ public class InMemoryUnfilteredPartitionIterator implements UnfilteredPartitionI
         public Row staticRow()
         {
             return partitionInfo.staticRow;
-        }
-
-        @Override
-        public DeletionTime partitionLevelDeletion()
-        {
-            return partitionInfo.partitionDeletion;
-        }
-
-        @Override
-        public EncodingStats stats()
-        {
-            return partitionInfo.encodingStats;
         }
     }
 }
